Move input handling utils out of amdllpc.cpp. NFC.
-  Add missing documentation and fix some existing comments.
-  Do not rely on global variables.
-  Use `llvm::StringLiteral` instead of C strings for file extensions.

Required for https://github.com/GPUOpen-Drivers/llpc/pull/1372.
